Move ExpressionStringifier to its own package
