+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_4/runtime/planDescription/PlanDescriptionArgumentSerializer.scala
@@ -20,10 +20,10 @@
 package org.neo4j.cypher.internal.compatibility.v3_4.runtime.planDescription
 
 import org.neo4j.cypher.internal.compatibility.v3_4.runtime.planDescription.InternalPlanDescription.Arguments._
-import org.neo4j.cypher.internal.frontend.v3_4.ExpressionStringifier
 import org.neo4j.cypher.internal.frontend.v3_4.helpers.UnNamedNameGenerator._
-import org.neo4j.cypher.internal.v3_4.expressions.SemanticDirection
+import org.neo4j.cypher.internal.frontend.v3_4.prettifier.ExpressionStringifier
 import org.neo4j.cypher.internal.v3_4.expressions
+import org.neo4j.cypher.internal.v3_4.expressions.SemanticDirection
 
 object PlanDescriptionArgumentSerializer {
   private val SEPARATOR = ", "
